fix buoyancy issue on distanceToSurface calculation
The distanceToSurface calculation seems to be wrong since
linkBoundingBox.Min().Z() is given in the world-frame and not in the
robot-frame as assumed by the previous calculation, explaining why the
robot pose was also necessary in the calculation.

Thus, the distance to surface can be calculated by simply subtracting
linkBoundingBox.Min().Z() from the water level.

This fixes the behavior of the robot to not respect the water level and
  not stop once being partially submerge when achieving the surface.